chore(sdk): Rename the `sliding_sync::view` module to `…::list`.
